:lipstick: Style header consistently with data tracker
